pyproject.toml
[tool.poetry]
name = "2023-autumn-ab-python-hw-14-team-dev-younggentech"
version = "0.1.0"
description = "Team work homework."
authors = ["Alexandr Ovsyannikov <alexov2004@gmail.com>", "Ilya Lystsov <lystsov.ilya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.23"
fastapi = "^0.106.0"
python-dotenv = "^1.0.0"
pyjwt = "^2.8.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
flake8 = "^6.1.0"
bandit = "^1.7.6"
black = "^23.12.1"
ruff = "^0.1.9"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"